feat: #2693 refactor tests
