+++ tests/brokers/base/consume.py
@@ -405,124 +405,6 @@ class BrokerConsumeTestcase(MultibrokerTestcase, BaseTestcaseConfig):
             with pytest.raises(AssertionError):
                 await subscriber.get_one(timeout=1e-24)
 
-    async def test_composition_context(
-        self,
-        queue: str,
-        mock: MagicMock,
-        event: asyncio.Event,
-    ) -> None:
-        broker = self.get_broker(
-            apply_types=True,
-            context=ContextRepo({"broker_context": "broker_context"}),
-        )
-
-        app = FastStream(broker, context=ContextRepo({"app_context": "app_context"}))
-
-        args, kwargs = self.get_subscriber_params(queue)
-
-        @broker.subscriber(*args, **kwargs)
-        async def handle(app_context=Context(), broker_context=Context()) -> None:
-            mock(app_context, broker_context)
-            event.set()
-
-        async with self.patch_broker(broker) as br, TestApp(app):
-            await asyncio.wait(
-                (
-                    asyncio.create_task(br.publish("", queue)),
-                    asyncio.create_task(event.wait()),
-                ),
-                timeout=self.timeout,
-            )
-
-            assert event.is_set()
-            mock.assert_called_once_with("app_context", "broker_context")
-
-    async def test_composition_context_merge(
-        self,
-        queue: str,
-        mock: MagicMock,
-        event: asyncio.Event,
-    ) -> None:
-        broker = self.get_broker(
-            apply_types=True,
-            context=ContextRepo({"context_var": "BROKER"}),
-        )
-
-        app = FastStream(broker, context=ContextRepo({"context_var": "APP"}))
-
-        args, kwargs = self.get_subscriber_params(queue)
-
-        @broker.subscriber(*args, **kwargs)
-        async def handle(context_var=Context()) -> None:
-            mock(context_var)
-            event.set()
-
-        async with self.patch_broker(broker) as br, TestApp(app):
-            await asyncio.wait(
-                (
-                    asyncio.create_task(br.publish("", queue)),
-                    asyncio.create_task(event.wait()),
-                ),
-                timeout=self.timeout,
-            )
-
-            assert event.is_set()
-            mock.assert_called_once_with("BROKER")
-
-    async def test_multi_broker_composition_context(
-        self,
-        queue: str,
-        mock: MagicMock,
-        mock2: MagicMock,
-        event: asyncio.Event,
-        event2: asyncio.Event,
-    ) -> None:
-        broker = self.get_broker(
-            apply_types=True,
-            context=ContextRepo({"broker_var": "BROKER 1"}),
-        )
-        broker2 = self.get_broker(
-            apply_types=True,
-            context=ContextRepo({"broker_var": "BROKER 2"}),
-        )
-
-        app = FastStream(broker, broker2, context=ContextRepo({"app_var": "APP"}))
-
-        args, kwargs = self.get_subscriber_params(queue)
-
-        @broker.subscriber(*args, **kwargs)
-        async def handle(app_var=Context(), broker_var=Context()) -> None:
-            event.set()
-            mock(app_var, broker_var)
-
-        args2, kwargs2 = self.get_subscriber_params(queue + "2")
-
-        @broker2.subscriber(*args2, **kwargs2)
-        async def handle2(app_var=Context(), broker_var=Context()) -> None:
-            event2.set()
-            mock2(app_var, broker_var)
-
-        async with (
-            self.patch_broker(broker) as br,
-            self.patch_broker(broker2) as br2,
-            TestApp(app),
-        ):
-            await asyncio.wait(
-                (
-                    asyncio.create_task(br.publish("", queue)),
-                    asyncio.create_task(br2.publish("", queue + "2")),
-                    asyncio.create_task(event.wait()),
-                    asyncio.create_task(event2.wait()),
-                ),
-                timeout=self.timeout,
-            )
-
-        assert event.is_set()
-        assert event2.is_set()
-        mock.assert_called_once_with("APP", "BROKER 1")
-        mock2.assert_called_once_with("APP", "BROKER 2")
-
-
 @pytest.mark.asyncio()
 class BrokerRealConsumeTestcase(BrokerConsumeTestcase):
     async def test_get_one(
